Moved serialization and TargetPicker into robotic, since it's only used there
