+++ pyobs/robotic/scheduler/targets/target.py
@@ -2,7 +2,7 @@ import abc
 from abc import ABCMeta
 from astropy.coordinates import SkyCoord
 
-from pyobs.utils.serialization import PolymorphicBaseModel
+from pyobs.robotic.utils.serialization import PolymorphicBaseModel
 from pyobs.utils.time import Time
